webApp/requirements.txt
pip==23.2.1
Flask==2.2.5
Werkzeug==2.2.3
jinja2==3.1.2
pycaret==3.0.0
hydra-core==1.3.2